Package Refactoring * re-structure file * add github repository link
