+++ index.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 const program = require('commander');
-const { crawlByThread } = require('./lib/crawler');
+const { crawlByThread } = require('./src/crawler');
 
 program.version("1.0.0").description("Lihkg Web Crawler")
